Adiciona modelo de busca no banco
